some small improvements
reading data from file and writing current hour added

+++ reactor.h
@@ -6,14 +6,14 @@ class Reactor {
 private:
     int number;
     int temperature;
-    int power; // MW
+    double power; // MW
 
 public:
     void measureTemp(); // writes out the current temperature of the reactor
     int getTemp(); // returns the temperature of the reactor
     int getNumber(); // returns the number of the reactor
-    int getPower(); // returns the power of the reactor
-    Reactor(int, int); // constructor
+    double getPower(); // returns the power of the reactor
+    Reactor(int, double); // constructor
 };
 
 #endif //PROJECT3_REACTOR_H
